MAJOR OVERHAUL part 5: Deprecating former devfile API
This was the OLD way of passing data to `ff_read_multi`, and it was a
mess. In all good code, the devfile argument was always null. Removing
the argument is not backwards compatible, so the function is now
`ff_read_frame_multi` (named for the underlying function,
`av_read_frame`). Use `ff_dev_file_waiting` to communicate with devices.