add more robust check for workers loaded as threads
Since emscripten-core/emscripten@37e1e68d51, the `pthread` worker.js file
no longer generated separately, and has instead been merged with the main
output javascript file, but the change makes all libav.js workers believe
they are the "main" worker and wait for a message that never arrives.

This makes it so that the worker name is checked and matched against
the newly introduced `self` property, and if it matches the expected
value, we know we are loaded as a thread.