include `PThread` in `EXPORTED_RUNTIME_METHODS`
Another change done in emscripten-core/emscripten@37e1e68d51 is
that `PThread` is not included in `EXPORTED_RUNTIME_METHODS` by
default anymore. This results in an exception being thrown when
initializing libav.js with `yesthreads` set to true.

This makes it so that it is once again included even when
building with newer emsdk versions.