More aggressive error detection for OOM
FFmpeg's write stack is not very resilient against OOM. Rather than
trying to change the writers in every single part of FFmpeg, I've just
made the underlying writeout function throw a JavaScript-level error if
it runs out of memory. This is of particular importance on iOS, where
memory is *very* limited.