Fixing CLI frontends *with* threads
... by not using threads. The I/O in libav.js is just not compatible
with true threads, because it needs to use Worker communication, so it
can't just be on a random worker. Realistically, thread performance is
going to come from the codecs and filters, not from the frontend, so the
fix is just to use emfiberthreads for the CLI even when pthreads are
available.